+++ lwb/metalang/cfg/cfg/src/main/java/mb/cfg/metalang/CfgEsvConfig.java
@@ -1,7 +1,6 @@
 package mb.cfg.metalang;
 
 import mb.cfg.CompileLanguageSpecificationShared;
-import mb.common.util.ListView;
 import mb.resource.hierarchical.ResourcePath;
 import mb.spoofax.compiler.language.StylerLanguageCompiler;
 import org.immutables.value.Value;
@@ -13,35 +12,15 @@ import java.io.Serializable;
  */
 @Value.Immutable
 public interface CfgEsvConfig extends Serializable {
-    class Builder extends ImmutableCfgEsvConfig.Builder {
-        public static ResourcePath getDefaultMainSourceDirectory(CompileLanguageSpecificationShared shared) {
-            return shared.languageProject().project().srcDirectory();
-        }
-
-        public static ResourcePath getDefaultMainFile(ResourcePath mainSourceDirectory) {
-            return mainSourceDirectory.appendRelativePath("main.esv");
-        }
-
-        public static boolean getDefaultIncludeLibSpoofax2Exports(CompileLanguageSpecificationShared shared) {
-            return shared.includeLibSpoofax2Exports();
-        }
-
-        public static ResourcePath getDefaultLibSpoofax2UnarchiveDirectory(CompileLanguageSpecificationShared shared) {
-            return shared.libSpoofax2UnarchiveDirectory();
-        }
-    }
+    class Builder extends ImmutableCfgEsvConfig.Builder {}
 
     static Builder builder() {return new Builder();}
 
 
     @Value.Default default CfgEsvSource source() {
-        final ResourcePath mainSourceDirectory = Builder.getDefaultMainSourceDirectory(compileLanguageShared());
-        return CfgEsvSource.files(
-            mainSourceDirectory,
-            Builder.getDefaultMainFile(mainSourceDirectory),
-            ListView.of(),
-            Builder.getDefaultIncludeLibSpoofax2Exports(compileLanguageShared()),
-            Builder.getDefaultLibSpoofax2UnarchiveDirectory(compileLanguageShared())
+        return CfgEsvSource.files(CfgEsvSource.Files.builder()
+            .compileLanguageShared(compileLanguageShared())
+            .build()
         );
     }