+++ lwb/metalang/cfg/cfg/src/main/java/mb/cfg/metalang/CfgStrategoSource.java
@@ -3,7 +3,6 @@ package mb.cfg.metalang;
 import mb.cfg.CompileLanguageSpecificationShared;
 import mb.common.util.ADT;
 import mb.resource.hierarchical.ResourcePath;
-import mb.spoofax.compiler.util.Shared;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.immutables.value.Value;
 
@@ -24,7 +23,7 @@ public abstract class CfgStrategoSource implements Serializable {
             }
         }
 
-        static Builder builder() { return new Builder(); }
+        static Builder builder() {return new Builder();}
 
 
         @Value.Default default ResourcePath mainSourceDirectory() {